// src/utils/interface/screens/productData.interface.ts
import React from "react";

export interface ProductDataProps {
  id?: number;
  customClassName?: string;
  productImage?: any;
  title?: string;
  price?: string;
  addToCart?: string;
  selectOption?: string;
  productLink?: string;
}

// export interface PriceRangeProps {
//   to: number;
//   form: number;
// }